guacozy_server/requirements-ldap.txt
-r requirements-base.txt
django-auth-ldap==4.5.0
whitenoise==6.6.0
python-ldap==3.4.4; sys_platform == 'linux'